separated enrollment_record from school
